refactor: extract search datafusion functions and plans
