+++ src/core/src/service/search/datafusion/udaf.rs
@@ -5,7 +5,7 @@
 // the Free Software Foundation, either version 3 of the License, or
 // (at your option) any later version.
 //
-// This program is distributed in the hope that it will be useful
+// This program is distributed in the hope that it will be useful,
 // but WITHOUT ANY WARRANTY; without even the implied warranty of
 // MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // GNU Affero General Public License for more details.
@@ -13,8 +13,4 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-pub use search::datafusion::plan::{deduplication, deduplication_exec};
-pub mod projections;
-#[cfg(feature = "vectorscan")]
-pub mod regex_projections;
-pub mod tantivy_optimize_exec;
+pub use search::datafusion::udaf::*;
