+++ src/search/src/datafusion/plan/deduplication_exec.rs
