+++ src/search/src/datafusion/udf/arrcount_udf.rs
