+++ src/search/src/datafusion/udf/arrindex_udf.rs
