+++ src/search/src/datafusion/udf/arrjoin_udf.rs
