+++ src/search/src/datafusion/udf/arrzip_udf.rs
