+++ src/search/src/datafusion/udf/match_all_udf.rs
@@ -20,7 +20,8 @@ use datafusion::{
     common::{DataFusionError, Result},
     logical_expr::{ColumnarValue, ScalarUDF, ScalarUDFImpl, Signature, Volatility},
 };
-pub use search::udf::{FUZZY_MATCH_ALL_UDF_NAME, MATCH_ALL_UDF_NAME};
+
+pub use crate::udf::{FUZZY_MATCH_ALL_UDF_NAME, MATCH_ALL_UDF_NAME};
 
 pub static MATCH_ALL_UDF: Lazy<ScalarUDF> = Lazy::new(|| ScalarUDF::from(MatchAllUdf::new()));
