+++ src/search/src/datafusion/udf/spath_udf.rs
