+++ src/search/src/datafusion/udf/str_match_udf.rs
