+++ src/search/src/lib.rs
@@ -13,6 +13,7 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+pub mod datafusion;
 pub mod enrichment;
 pub mod sql;
 pub mod udf;
